backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
aio-pika==9.4.2
qdrant-client==1.11.3
python-dotenv==1.0.1
sentence-transformers==3.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.2.0
